constitution-chatbot-starter/requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
PyPDF2
faiss-cpu
python-dotenv
